refactor(api): neutralize web push module names
