+++ turbo/apps/api/src/signals/routes/push-subscriptions.ts
@@ -4,7 +4,7 @@ import { pushSubscriptionsContract } from "@okouai/api-contracts/contracts/push-
 import { authContext$ } from "../auth/auth-context";
 import { authRoute } from "../auth/auth-route";
 import { bodyResultOf } from "../context/request";
-import { registerPushSubscription$ } from "../services/zero-push-subscriptions.service";
+import { registerPushSubscription$ } from "../services/push-subscriptions.service";
 import type { RouteEntry } from "../route-entry";
 
 const registerInner$ = command(async ({ get, set }, signal: AbortSignal) => {
@@ -34,7 +34,7 @@ const registerInner$ = command(async ({ get, set }, signal: AbortSignal) => {
   return { status: 201 as const, body: { success: true as const } };
 });
 
-export const zeroPushSubscriptionsRoutes: readonly RouteEntry[] = [
+export const pushSubscriptionsRoutes: readonly RouteEntry[] = [
   {
     route: pushSubscriptionsContract.register,
     handler: authRoute({}, registerInner$),
